Move ResponseGenericDTO to 'response' folder
