+++ Server/server_ed_platform/src/main/java/com/nocountry/server_ed_platform/dtos/Response/ResponseGenericDTO.java
@@ -1,4 +1,4 @@
-package com.nocountry.server_ed_platform.dtos;
+package com.nocountry.server_ed_platform.dtos.Response;
 
 
 import lombok.AllArgsConstructor;
